implement merging of compatible structs
Often a disjunction without a schema is of the form:

	{a!: string, b?: int} | {c!: string, b: int}

where even though the structs aren't easily discriminatable,
all the fields are compatible; that is, if our aim is to use some
concrete type representation (e.g. a Go type), we can
potentially use a single Go type to hold all arms of the
disjunction.

We implement support for this so if MergeCompatible
is enabled, we will merge all compatible structs
as we do compatible atoms.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>